requirements.txt
django==4.2.15              # Framework principal
djangorestframework         # API REST pour Django
django-cors-headers         # Gestion des CORS
channels                    # WebSockets avec Django
channels-redis              # Backend Redis pour Channels
daphne                      # Serveur ASGI
psycopg2-binary             # Connecteur PostgreSQL
python-dotenv               # Chargement des variables d'environnement
# ccxt                      # SUPPRIMÉ - Remplacé par clients natifs
redis>=4.3.4                # Client Redis avec support asyncio
websockets                  # Support WebSocket pur Python
pandas-ta-classic           # Indicateurs techniques (trading)
cryptography                # Chiffrement des clés API (Fernet)
aiohttp==3.9.1          # Serveur HTTP asynchrone léger
aioredis==2.0.1         # Client Redis asynchrone
uvloop==0.19.0          # Event loop optimisé (optionnel mais recommandé)
loguru                  # Logging structuré JSON pour tous les terminaux
